+++ src/components/bluepad32/uni_hid_device.c
@@ -125,12 +125,6 @@ uni_hid_device_t* uni_hid_device_get_instance_for_connection_handle(hci_con_hand
     return NULL;
 }
 
-uni_hid_device_t* uni_hid_device_get_instance_for_idx(int idx) {
-    if (idx < 0 || idx >= CONFIG_BLUEPAD32_MAX_DEVICES)
-        return NULL;
-    return &g_devices[idx];
-}
-
 uni_hid_device_t* uni_hid_device_get_instance_with_predicate(uni_hid_device_predicate_t predicate, void* data) {
     for (int i = 0; i < CONFIG_BLUEPAD32_MAX_DEVICES; i++) {
         // Only "ready" devices are propagated
@@ -140,6 +134,20 @@ uni_hid_device_t* uni_hid_device_get_instance_with_predicate(uni_hid_device_pred
             return &g_devices[i];
     }
     return NULL;
+}
+
+uni_hid_device_t* uni_hid_device_get_instance_for_idx(int idx) {
+    if (idx < 0 || idx >= CONFIG_BLUEPAD32_MAX_DEVICES)
+        return NULL;
+    return &g_devices[idx];
+}
+
+int uni_hid_device_get_idx_for_instance(uni_hid_device_t* d) {
+    int idx = (d - &g_devices[0]) / sizeof(g_devices[0]);
+
+    if (idx < 0 || idx >= CONFIG_BLUEPAD32_MAX_DEVICES)
+        return -1;
+    return idx;
 }
 
 uni_hid_device_t* uni_hid_device_get_first_device_with_state(uni_bt_conn_state_t state) {